0412/project-23/requirements.txt
apache-airflow==2.8.1
beautifulsoup4
